crawler/requirements.txt
# YouTube 데이터 수집
google-api-python-client>=2.100.0
youtube-transcript-api==0.6.1  # 0.6.1 고정 (1.x 버전은 API 변경됨)

# AI 분석
openai>=1.0.0
google-generativeai>=0.3.0

# 웹 크롤링 (매장 데이터)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0
aiohttp>=3.9.0

# 데이터 처리
pandas>=2.0.0
python-dotenv>=1.0.0

# 데이터베이스
sqlalchemy>=2.0.0
supabase>=2.0.0  # 클라우드 DB

# 관리자 대시보드
streamlit>=1.28.0

# 스케줄링
schedule>=1.2.0

# 비동기 처리
asyncio>=3.4.3
ratelimit>=2.2.1

# 유틸리티
tqdm>=4.66.0
